trydjango/requirements.txt
djangorestframework
Django==2.1.3
Pillow==5.3.0
pytz==2018.7
